[scanner] fix: refactor oversized source files (codeGenerator.templates, Marketplace)
Fixes #20330

Refactors the 2 oversized production files:

1. codeGenerator.templates.ts (1326→324 lines):
   - Split massive generateCardRenderFunction switch into 7 category modules
   - templates.cluster.ts - cluster health, metrics, workload, namespace
   - templates.pods.ts - pod issues, top pods, app status
   - templates.events.ts - event summary, warning events, alerts
   - templates.storage.ts - storage, PVC, network, services
   - templates.operators.ts - operators, helm, opencost, GPU
   - templates.github.ts - all GitHub/CI cards, security, providers
   - templates.consoleai.ts - Console AI cards
   - templates.shared.ts - shared template constants

2. Marketplace.tsx (1051→506 lines):
   - Extracted CNCFProgressBanner component
   - Extracted MarketplaceCard component
   - Extracted MarketplaceRow component
   - Extracted AuthorBadge and DifficultyBadge to MarketplaceBadges
   - Moved constants to marketplaceConstants.ts

Note: 18 test files over 500 lines were intentionally excluded — large test files are normal and splitting them reduces readability.
